新文件：   02_0address.c 修改：     02_container_of.c 修改：     02_typeof.c 新文件：   02_typeof_1.c

+++ CppTest/Linux_Kernel_C/02_typeof.c
@@ -96,8 +96,8 @@ void test04()
 {
 	int a = 4;
 	float b = 5.6;
-	int c = max(a, b);
-	printf("%d\n", c);
+	float c = max(a, b);
+	printf("%f\n", c);
 	printf("---------------------------\n");
 }
